Update resolution_and_sharpness_of_images.py
Created a loop that analyzes each image and saves the analysis as an image with the same name as the original but with _mtf.png appended instead of .png. These also each are analyzed and saved in order of the original images, and in the same folder as the original images. So each one is saved right next to the image it analyzed.